backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
Pillow==12.0.0
python-decouple==3.8
django-cors-headers==4.9.0
psycopg2-binary==2.9.9
redis==5.1.1
celery==5.4.0
django-extensions==3.2.3
